Vagrantfile: Override default arguments to rsync command
The default arguments to rsync in Vagrant are "--verbose --archive --delete -z
--copy-links". With the current ceph-ansible directory structure this is
problematic as the symlink ceph.ceph-common ends up being copied as a
directory, thus creating multiple copies of the files beneath it. We modify one
copy of the file but ceph-ansible refers to the file via its symlink and
therefore gets the unmodified copy. Overriding the defaul arguments should
resolve this.

Signed-off-by: Brad Hubbard <bhubbard@redhat.com>